requirements.txt
streamlit
elevenlabs
pydub
python-docx
PyPDF2
python-dotenv
ffmpeg-python
nltk
openai==1.106.1
boto3
tenacity